Use updated EA framework and types
