+++ packages/sources/cryptocompare-test/package.json
@@ -28,7 +28,7 @@
     "start": "yarn server:dist"
   },
   "dependencies": {
-    "@chainlink/external-adapter-framework": "0.5.1",
+    "@chainlink/external-adapter-framework": "0.10.0",
     "tslib": "^2.3.1"
   },
   "devDependencies": {
