+++ packages/sources/cryptocompare-test/src/config/index.ts
@@ -4,8 +4,7 @@ export const DEFAULT_WS_API_ENDPOINT = 'wss://streamer.cryptocompare.com/v2'
 export const defaultEndpoint = 'crypto'
 
 export const customSettings = {
-  // TODO: Temp commented out to not cause conflicts with EA framework config. Uncomment on the next EA framework release
-  /*API_KEY: {
+  API_KEY: {
     description: 'The CryptoCompare API key',
     type: 'string',
     required: true,
@@ -15,5 +14,5 @@ export const customSettings = {
     description: 'The websocket API key to authenticate with, if different from API_KEY',
     type: 'string',
     sensitive: true,
-  },*/
+  },
 } as const
